Implimented changes
