+++ base_decorator.rb
@@ -1,6 +1,8 @@
 require './namable_class'
 
 class BaseDecorator < Nameable
+  attr_accessor :nameable
+
   def initialize(nameable)
     @nameable = nameable
     super()
@@ -8,18 +10,5 @@ class BaseDecorator < Nameable
 
   def correct_name
     @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    name = @nameable.correct_name.capitalize
-    name.length > 10 ? name[0..9] : name
   end
 end
